feat: integrate structured data directly into LocaleLayout for improved SEO and remove Head component
