+++ src/app/[locale]/_components/Home.tsx
@@ -51,11 +51,11 @@ const Home = async () => {
         className={merge(
           "root relative",
           "h-[calc(100dvh-52px)]",
-          "sm:h-[calc(100dvh-60px-277px)]",
-          "md:h-[calc(100dvh-60px-309px)]",
-          "lg:h-[calc(100dvh-60px-231px)]",
-          "xl:h-[calc(100dvh-60px-232px)]",
-          "2xl:h-[calc(100dvh-60px-165px)]",
+          "sm:h-[calc(100dvh-60px-245px)]",
+          "md:h-[calc(100dvh-60px-279px)]",
+          "lg:h-[calc(100dvh-60px-199px)]",
+          "xl:h-[calc(100dvh-60px-201px)]",
+          "2xl:h-[calc(100dvh-60px-132px)]",
           "flex justify-center md:justify-start gap-5"
         )}
       >
